requirements.txt
django==4.2.6
djangorestframework==3.14.0
mysqlclient==2.2.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
openpyxl==2.6.0